FIXED Render loop if statement not working as intended. The memory leak still exists but is used so rarely that it becomes near irrelevant. All my "creates"
and allocates are working as intended and released as intended. Ive honestly 0 clue of the cause of it. I want to blame FW1Wrapper but I simply do not know.

The memory leak only occurs when we edit the strings. Every time there are a few bytes lost. If we update a string every frame, it takes up 1 MB every 9 seconds.
But now with reduced updates of the string, it takes a very long time to do so.

CHANGED Font wrapper is now shared by all fonts rather than each having their own. Should increase performance somewhat.