RENAMED 2-3 variable names.
RENAMED MENU inside the controls.json

This task was unneccessary in that we have the functionality we want. Even though we reuse code in the controls.json, it is as intended. It have way
more functionality than grouping everything below 1 node. If we want to export it "to the user like in a real game". We would open up the PLAY
controls for them. The rest should not be handled by the user.