FIXED SoundModule data structure.
ADDED example for adding and playing sound in MenuState. Activates when play button is pressed.

CHANGED  all state constructors to also add the SoundModule pointer to class.

ADDED SoundModule in Game.cpp and a pointer to that object in every state. This is not optimal but has to be done in our current state setup.